add graphAnalysis LoadData ,RenameTab,QueryLoadStatus case

+++ src/main/java/com/zjlab/qa/tests/graphAnalysis/DeleteGraphByIdTest.java
@@ -8,6 +8,8 @@ import com.zjlab.qa.utils.GetJsonValueUtil;
 import com.zjlab.qa.utils.ReadExcelUtil;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.util.EntityUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 import org.testng.Reporter;
 import org.testng.annotations.AfterClass;
@@ -17,15 +19,17 @@ import org.testng.annotations.Test;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class DeleteGraphByIdTest {
+    private static final Logger log= LoggerFactory.getLogger(DeleteGraphByIdTest.class);
     private GraphAnalysisClientApi graphAnalysisClient;
     private List<Map<String, String>> deleteGraphData;
-    private List<String> projectGraphIds;
+    private List<String> proIds;
     private ProjectManage projectManage;
-    private String projId;
+    private String proId;
     private String graphId;
 
 
@@ -33,13 +37,13 @@ public class DeleteGraphByIdTest {
     public void setUp(){
         projectManage=new ProjectManage();
         graphAnalysisClient=new GraphAnalysisClientApi();
-        projectGraphIds=new ArrayList<String>();
+        proIds=new ArrayList<String>();
         deleteGraphData = ReadExcelUtil.getExcuteList("deleteGraphById-test.xlsx");
 
 
 
     }
-//    通过读取Excel获取测试数据入参
+//    通过读取Excel获取测试数据Request Parameter
     @DataProvider
     public Object[][] deleteGraphData(){
         Object[][] files = new Object[deleteGraphData.size()][];
@@ -54,38 +58,44 @@ public class DeleteGraphByIdTest {
         String params = (String) param.get("params");
         String expectCode = (String) param.get("expectCode");
         String expectMessage = (String) param.get("expectMessage");
-        List<String> placeholders = ParseKeyword.getKeywords(params);
+        String isRun = (String) param.get("isRun");
+        if(isRun.contains("1")) {
+            List<String> placeholders = ParseKeyword.getKeywords(params);
 //替换Excel中通过$占位的参数
-        if(placeholders.size()>0 && placeholders.contains("projectId") && placeholders.contains("graphId")){
+            if (placeholders.size() > 0 && placeholders.contains("projectId") && placeholders.contains("graphId")) {
+                Map<String, String> map = new HashMap<String, String>();
 //            新建项目，获取项目id
-            CloseableHttpResponse projRe= projectManage.create();
-            String responseStr = EntityUtils.toString(projRe.getEntity(), "UTF-8");
-            JSONObject responseJson = JSONObject.parseObject(responseStr);
-            projId = GetJsonValueUtil.getValueByJpath(responseJson, "result");
-            String addGraph="{\"projectId\":"+projId+"}";
+                CloseableHttpResponse projRe = projectManage.create();
+                String responseStr = EntityUtils.toString(projRe.getEntity(), "UTF-8");
+                JSONObject responseJson = JSONObject.parseObject(responseStr);
+                proId = GetJsonValueUtil.getValueByJpath(responseJson, "result");
+                proIds.add(proId);
+                String addGraph = "{\"projectId\":" + proId + "}";
 //        新建标签页，获取标签页ID
-            CloseableHttpResponse graphRe= graphAnalysisClient.addGraph(addGraph);
-            String graphReStr = EntityUtils.toString(graphRe.getEntity(), "UTF-8");
-            JSONObject graphReJson = JSONObject.parseObject(graphReStr);
-            graphId = GetJsonValueUtil.getValueByJpath(graphReJson, "result/id");
-            params="{\"projectId\":"+projId+",\"graphId\":"+graphId+"}";
+                CloseableHttpResponse graphRe = graphAnalysisClient.addGraph(addGraph);
+                String graphReStr = EntityUtils.toString(graphRe.getEntity(), "UTF-8");
+                JSONObject graphReJson = JSONObject.parseObject(graphReStr);
+                graphId = GetJsonValueUtil.getValueByJpath(graphReJson, "result/id");
+                map.put("projectId", proId);
+                map.put("graphId", graphId);
+                params = ParseKeyword.replacePeso(params, map);
+//            params="{\"projectId\":"+proId+",\"graphId\":"+graphId+"}";
+            }
+
+            CloseableHttpResponse re = graphAnalysisClient.deleteGraphById(params);
+
+            //获取响应内容
+            String delGraphString = EntityUtils.toString(re.getEntity(), "UTF-8");
+            log.info("Request URL：" + graphAnalysisClient.getUrl() + "，Request Parameter：" + params);
+            log.info("Response：" + delGraphString);
+            //创建JSON对象  把得到的响应字符串 序列化成json对象
+            JSONObject delGraphJson = JSONObject.parseObject(delGraphString);
+            String code = GetJsonValueUtil.getValueByJpath(delGraphJson, "code");
+            String message = GetJsonValueUtil.getValueByJpath(delGraphJson, "message");
+
+            Assert.assertEquals(code, expectCode, title + "; 实际的code：" + code + "，期望返回的code：" + expectCode);
+            Assert.assertTrue(message.contains(expectMessage), title + "; 实际的message：" + message + "，期望返回的message：" + expectMessage);
         }
-
-        CloseableHttpResponse re = graphAnalysisClient.deleteGraphById(params);
-
-        //获取响应内容
-        String delGraphString = EntityUtils.toString(re.getEntity(), "UTF-8");
-        System.out.println("接口请求"+graphAnalysisClient.getUrl()+"，入参："+params);
-        System.out.println("接口返回："+delGraphString);
-        //创建JSON对象  把得到的响应字符串 序列化成json对象
-        JSONObject delGraphJson = JSONObject.parseObject(delGraphString);
-        String code = GetJsonValueUtil.getValueByJpath(delGraphJson, "code");
-        String message = GetJsonValueUtil.getValueByJpath(delGraphJson, "message");
-
-        Assert.assertEquals(code,expectCode,title+"; 实际的code："+code+"，期望返回的code："+expectCode);
-        boolean isMessage=message.contains(expectMessage);
-        Assert.assertTrue(message.contains(expectMessage),title+"; 实际的message："+message+"，期望返回的message："+expectMessage);
-
 
     }
 
@@ -94,21 +104,25 @@ public class DeleteGraphByIdTest {
      */
     @AfterClass
     public void tearDown(){
-        String delPrams="{\"id\":"+projId+"}";
-        CloseableHttpResponse re= projectManage.deleteById(delPrams);
-        String responseString = null;
-        try {
-            responseString = EntityUtils.toString(re.getEntity(), "UTF-8");
-        } catch (IOException e) {
-            e.printStackTrace();
+        for (String proId:proIds
+        ) {
+            String delPrams="{\"id\":"+proId+"}";
+            CloseableHttpResponse re= projectManage.deleteById(delPrams);
+            String responseString = null;
+            try {
+                responseString = EntityUtils.toString(re.getEntity(), "UTF-8");
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+            log.info("############################################# Start Clear Test Data ##############################################");
+            log.info("Request URL："+projectManage.getUrl()+"，Request Parameter："+delPrams);
+            log.info("Response："+responseString);
+            log.info("########################################## Clear Test Data Success ##############################################");
+
         }
 
-        System.out.println("##############开始删除测试新增的项目##################");
-        System.out.println("接口请求"+projectManage.getUrl()+"，入参："+delPrams);
-        System.out.println("接口返回："+responseString);
-        System.out.println("##############成功删除测试新增的项目##################");
-
-        }
+    }
 
 }
